+++ base-build.sh
@@ -7,4 +7,4 @@ go mod vendor
 [ -e internal/gen ] || mkdir -p internal/gen
 protoc -Iproto -I/usr/include --plugin=$GOPATH/bin/protoc-gen-go --go_out=./internal/gen --python_out=./assets/response_catcher proto/reports.proto
 protoc -Iproto -I/usr/include --plugin=$GOPATH/bin/protoc-gen-go --go_out=./internal/gen proto/tensorflow_serving/config/*.proto
-(go run build.go -r -dirs=tools/serving-bridge,tools/queue-status,internal,cmd) 2>&1 | tee "$RUNNER_BUILD_LOG"
+(go run build.go -r -dirs=tools/serving-bridge,tools/queue-scaler,internal,cmd) 2>&1 | tee "$RUNNER_BUILD_LOG"