+++ cmd/runner/main.go
@@ -20,9 +20,12 @@ import (
 	"github.com/leaf-ai/go-service/pkg/server"
 
 	aws_ext "github.com/leaf-ai/studio-go-runner/internal/aws"
+
+	"github.com/leaf-ai/studio-go-runner/internal/cpu_resource"
 	"github.com/leaf-ai/studio-go-runner/internal/cuda"
 	"github.com/leaf-ai/studio-go-runner/internal/defense"
-	"github.com/leaf-ai/studio-go-runner/internal/runner"
+	"github.com/leaf-ai/studio-go-runner/internal/disk_resource"
+	"github.com/leaf-ai/studio-go-runner/internal/resources"
 
 	"github.com/davecgh/go-spew/spew"
 
@@ -313,10 +316,10 @@ func validateResourceOpts() (errs []kv.Error) {
 		errs = append(errs, kv.Wrap(err).With("stack", stack.Trace().TrimRuntime()))
 	}
 
-	if err = runner.SetCPULimits(limitCores, limitMem); err != nil {
+	if err = cpu_resource.SetCPULimits(limitCores, limitMem); err != nil {
 		errs = append(errs, kv.Wrap(err, "the cores, or memory limits on command line option were invalid").With("stack", stack.Trace().TrimRuntime()))
 	}
-	avail, err := runner.SetDiskLimits(*tempOpt, limitDisk)
+	avail, err := disk_resource.SetDiskLimits(*tempOpt, limitDisk)
 	if err != nil {
 		errs = append(errs, kv.Wrap(err, "the disk storage limits on command line option were invalid").With("stack", stack.Trace().TrimRuntime()))
 	} else {
@@ -460,7 +463,7 @@ func startServices(ctx context.Context, cancel context.CancelFunc, statusC chan 
 
 	// loops doing prometheus exports for resource consumption statistics etc
 	// on a regular basis
-	server.StartPrometheusExporter(ctx, *promAddrOpt, &runner.Resources{}, time.Duration(10*time.Second), logger)
+	server.StartPrometheusExporter(ctx, *promAddrOpt, &resources.Resources{}, time.Duration(10*time.Second), logger)
 
 	// The timing for queues being refreshed should me much more frequent when testing
 	// is being done to allow short lived resources such as queues etc to be refreshed